test : Add E2E tests for Helm Install and Uninstall functionality
Add e2e tests for these goals:
- `k8s:helm-install`
- `k8s:helm-uninstall`
- `oc:helm-install`
- `oc:helm-uninstall`
